PCHR-3162: Resolve merge conflict
After an upgrader was added we need to shift the epic upgrader steps up
to avoid a conflict. The 1020 upgrader was copied from Tunbola's commit
0dbab999255c051cfcd040036263bc47757140e0
