+++ uk.co.compucorp.civicrm.hrleaveandabsences/CRM/HRLeaveAndAbsences/Upgrader.php
@@ -30,6 +30,7 @@ class CRM_HRLeaveAndAbsences_Upgrader extends CRM_HRLeaveAndAbsences_Upgrader_Ba
   use CRM_HRLeaveAndAbsences_Upgrader_Step_1022;
   use CRM_HRLeaveAndAbsences_Upgrader_Step_1023;
   use CRM_HRLeaveAndAbsences_Upgrader_Step_1024;
+  use CRM_HRLeaveAndAbsences_Upgrader_Step_1025;
 
   /**
    * A list of directories to be scanned for XML installation files
