+++ uk.co.compucorp.civicrm.hrleaveandabsences/CRM/HRLeaveAndAbsences/Upgrader/Step/1022.php
@@ -3,22 +3,18 @@
 trait CRM_HRLeaveAndAbsences_Upgrader_Step_1022 {
   
   /**
-   * Default existing work pattern to As per contract (new default)
+   * Create new option group hrleaveandabsences_work_pattern_change_reason
+   * with values As per contract, Change in contractual hours and
+   * Change in contract type.
    *
    * @return bool
    */
-  public function upgrade_1022()
-  {
-    $query = 'UPDATE civicrm_hrleaveandabsences_contact_work_pattern SET change_reason =
-      (
-        SELECT cov.value FROM civicrm_option_value cov LEFT JOIN civicrm_option_group cog
-          ON (cog.id = cov.option_group_id)
-          WHERE cog.name = "hrleaveandabsences_work_pattern_change_reason"
-          AND cov.is_default = 1
-      )';
-    
-    CRM_Core_DAO::executeQuery($query);
-    
+  public function upgrade_1022() {
+    $import = new CRM_Utils_Migrate_Import();
+    $extension = 'uk.co.compucorp.civicrm.hrleaveandabsences';
+    $file = 'xml/option_groups/work_pattern_change_reason_install.xml';
+  
+    $import->run(CRM_Core_Resources::singleton()->getPath($extension, $file));
     return TRUE;
   }
 }
